requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
backtrader==1.9.78.123
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
Crawl4AI==0.4.248
cryptography==44.0.1
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
display==1.0.0
distro==1.9.0
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.0.3
filelock==3.17.0
filetype==1.2.0
fonttools==4.55.8
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
genv==1.4.3
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
hn-sdk==0.1.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.18
langchain-anthropic==0.3.4
langchain-community==0.3.17
langchain-core==0.3.34
langchain-google-genai==2.0.9
langchain-openai==0.3.2
langchain-text-splitters==0.3.6
langgraph==0.2.68
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.2
litellm==1.61.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
openai==1.63.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pillow==10.4.0
platformdirs==4.3.6
playwright==1.50.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.5.0
tenacity==9.0.0
tf-playwright-stealth==1.1.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.52
zipp==3.21.0
zstandard==0.23.0